Fetch full repository on GitHub Actions workflow
Fetching with the default fetch-depth of 1 led to a strange bug where the
commit fetched would be different from the one that triggered the build, unless
the build was triggered manually. It is not entirely clear why this happens,
and debugging with using things like `git rev-parse HEAD` etc didn't give any
hints. The issue [here](https://github.com/actions/checkout/issues/439) seems
to be the problem, and this commit addresses it by setting fetch-depth to 0.