Change names for one dataset for graph500 benchmarks
The graph500 benchmarks don't follow the convention required by sandmark
nightly to be able to normalize the multicore benchmarks using the sequential
benchmark data. This commit changes the data for one benchmark run so that the
data can be visualized. A separate PR to ocaml-bench/sandmark will fix the
names of the benchmark for future runs.