validate_run: Relax constraint for marking bench run as successful
Previously, any run with any errors in the log file or without any benchmark
data were marked as failures.  This commit relaxes the constraint and marks
runs with even one successful benchmark as success.

Also, in the status page we display the number of benchmarks that successfully
run in the total number of benchmarks configured to run.