Update Streamlit to latest version
We were using an outdated version of Streamlit 1.14.0 and this commit updates
to the latest version 1.22.0. This makes it easier to deploy any PR branches of
the app to Streamlit Cloud.

With the older version, we run into a bug in the Streamlit deployment where the
deploy process tries to be smart about "downgrading" Altair to a compatible
version. I'm not sure why a different version of Altair is installed in the
first place, and why it is later downgraded. This results in different versions
of some packages being installed than those specified in our requirements.txt.